* include/spells.h     - The ASPELL macro now defines it's level parameter as an int       instead of a byte since it's calculated and passed into       call_magic() as an int.  Maybe imms who want to have fun by       setting themselves to obscene gens and trying out spells       will stop reporting bugs to me now...

+++ src/include/spells.h
@@ -862,7 +862,7 @@ struct gun_hit_type {
 
 
 #define ASPELL(spellname) \
-	void spellname(__attribute__ ((unused)) byte level, \
+	void spellname(__attribute__ ((unused)) int level, \
 		__attribute__ ((unused)) struct Creature *ch, \
 		__attribute__ ((unused))  struct Creature *victim, \
 		__attribute__ ((unused)) struct obj_data *obj, \
